feat: add CriteriaEvaluation and TestCaseEvaluation components for enhanced submission evaluation
